pyproject.toml
[project]
name = "livrable-p9"
version = "0.1.0"
description = ""
authors = [
    {name = "S-S-Zheng",email = "zheng.shipo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpy (>=2.4.2,<3.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "tenacity (>=9.1.4,<10.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-core (>=1.2.17,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-mistralai (>=1.1.1,<2.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
]

[tool.poetry]
packages = [{include = "livrable_p9", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "notebook (>=7.5.4,<8.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "ipywidgets (>=8.1.8,<9.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "ragas (>=0.4.3,<0.5.0)",
    "datasets (>=4.6.1,<5.0.0)",
    "ruff (>=0.15.5,<0.16.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]

[tool.ruff]
line-length = 100
target-version = "py312" # ou votre version de python

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B"]
ignore = [
    "E203",
    "B008", # Autorise les appels de fonctions dans les arguments (FastAPI Depends)
    # "E501", # Optionnel : Ignore les lignes trop longues si tu préfères les gérer manuellement
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B017"] # Autorise pytest.raises(Exception) dans les tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"